Refactored the buildOptionsFromAddonCommands util function to support the current changes and created a different one for the headless menu usecase.
