+++ src/components/Editor/Menu/Headless/utils.js
@@ -163,6 +163,16 @@ const createMenuOptions = ({
   },
 ];
 
+export const buildOptionsFromAddonCommands = ({ editor, commands }) => {
+  const { to } = editor.state.selection;
+
+  return commands.map(option => ({
+    ...option,
+    active: option.active?.({ editor }),
+    command: () => option.command?.({ editor, range: { from: to, to } }),
+  }));
+};
+
 export const buildMenuOptions = ({
   tooltips,
   editor,
